refactor<>: Consistencia DDC/DER/MYSQL/CODIGO

+++ src/main/java/domain/measurements/ActivityData.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class ActivityData {
 
   private ConsumptionType consumptionType;
-  private Integer consumptionValue;
+  private Double consumptionValue;
   private PeriodicityFormat periodicityFormat;
   private String periodicity;
